requirements.txt
# Core dependencies
flask==3.1.0
langchain==0.3.23
langchain-openai==0.3.14
langchain-deepseek==0.0.1
langgraph==0.3.31
python-dotenv==1.0.1

# Additional dependencies
requests==2.32.3
pydantic==2.11.3
